requirements.txt
streamlit
google-generativeai
python-dotenv
nbformat
pandas
numpy
matplotlib
seaborn
markdown
plotly.express